Improvements: provide modularity. Clean-up from "console.log" for success test.
